refactor: Move and rename files for better organization
- Moved files to different folders to improve project structure.
- Renamed files to adhere to new naming conventions.
