+++ src/Wigo.Service/Handlers/AddTopUpTransactionHandler.cs
@@ -12,18 +12,18 @@ public class AddTopUpTransactionHandler : IRequestHandler<AddTopUpTransactionCom
     private readonly IUserRepository _userRepository;
     private readonly IBeneficiaryRepository _beneficiaryRepository;
     private readonly ITopUpTransactionRepository _topUpTransactionRepository;
-    private readonly IExternalAccountBalanceService _accountBalanceService;
+    private readonly IExternalAccountBalanceService _externalAccountBalanceService;
 
     public AddTopUpTransactionHandler(
         IUserRepository userRepository,
         IBeneficiaryRepository beneficiaryRepository,
         ITopUpTransactionRepository topUpTransactionRepository,
-        IExternalAccountBalanceService accountBalanceService)
+        IExternalAccountBalanceService externalAccountBalanceService)
     {
         _userRepository = userRepository;
         _beneficiaryRepository = beneficiaryRepository;
         _topUpTransactionRepository = topUpTransactionRepository;
-        _accountBalanceService = accountBalanceService;
+        _externalAccountBalanceService = externalAccountBalanceService;
     }
 
     public async Task<ServiceResult<Guid>> Handle(AddTopUpTransactionCommand command, CancellationToken cancellationToken)
@@ -44,7 +44,7 @@ public class AddTopUpTransactionHandler : IRequestHandler<AddTopUpTransactionCom
         }
 
         // Check user's balance
-        var balance = await _accountBalanceService.GetAccountBalanceAsync(command.UserAccountBalanceNumber);
+        var balance = await _externalAccountBalanceService.GetAccountBalanceAsync(command.UserAccountBalanceNumber);
         var totalAmount = command.Amount;
         if (totalAmount > balance)
         {
@@ -68,7 +68,7 @@ public class AddTopUpTransactionHandler : IRequestHandler<AddTopUpTransactionCom
         }
 
         // Debit user's balance through external service
-        var debitResult = await _accountBalanceService.DebitAccountBalanceAsync(command.UserAccountBalanceNumber, totalAmount);
+        var debitResult = await _externalAccountBalanceService.DebitAccountBalanceAsync(command.UserAccountBalanceNumber, totalAmount);
         if (!debitResult)
         {
             return ServiceResult<Guid>.FailureResult("Failed to debit balance.");
